% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add_matrices_bycol}
\alias{add_matrices_bycol}
\title{sum of two matrices by column}
\usage{
add_matrices_bycol(x, y)
}
\arguments{
\item{x}{a matrix}

\item{y}{a matrix with the same dimensions}
}
\value{
a matrix
}
\description{
sum of two matrices by column
}
\keyword{internal}
