% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp.R
\name{adj_bg_color}
\alias{adj_bg_color}
\title{Adjust the background of plot}
\usage{
adj_bg_color(plot, light, input)
}
\arguments{
\item{plot}{The plot (plotly) to adjust}

\item{input}{The shiny app input object}

\item{ligght}{The actual theme used}
}
\description{
Adjust the background of plot (light or dark mode)
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
