% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembling.R
\name{belongsFCM}
\alias{belongsFCM}
\title{membership matrix calculator for FCM algorithm}
\usage{
belongsFCM(data, centers, dots)
}
\arguments{
\item{data}{a matrix (the dataset used for clustering)}

\item{centers}{a matrix (the centers of the clusters)}

\item{dots}{a list of other arguments specific to FCM}
}
\value{
a matrix with the new membership values
}
\description{
membership matrix calculator for FCM algorithm
}
\keyword{internal}
