% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostrap_clust_validation.R
\name{boot_worker}
\alias{boot_worker}
\title{Bootstrap check the robustness of a classification (multicore)}
\usage{
boot_worker(object, wdata, tol, maxiter, init)
}
\arguments{
\item{object}{A FCMres object, typically obtained from functions CMeans, GCMeans, SFCMeans, SGFCMeans}

\item{wdata}{The lagged dataset if necessary, can be NULL if not required}

\item{tol}{The tolerance criterion used in the evaluateMatrices function for
convergence assessment}

\item{maxiter}{An integer for the maximum number of iteration}

\item{init}{A string indicating how the initial centres must be selected. "random"
indicates that random observations are used as centres. "kpp" use a distance based method
resulting in more dispersed centres at the beginning. Both of them are heuristic.}
}
\value{
A list, similar to a FCMres object, but with only necessary slots for cluster bootstraping.
}
\description{
Worker function for cluster bootstrapping
}
\details{
The worker function for the functions boot_group_validation and boot_group_validation.mc
}
\examples{
# this is an internal function, no example provided
}
\keyword{internal}
