% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp.R
\name{boxplots_ui}
\alias{boxplots_ui}
\title{Set the box plot pannel}
\usage{
boxplots_ui(n)
}
\arguments{
\item{n}{The number of plots}
}
\description{
A simple function to prepare the box plot pannel in uncertainty
analysis
}
\examples{
#This is an internal function, no example provided
}
\keyword{interal}
