% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostrap_clust_validation.R
\name{calc_jaccard_idx}
\alias{calc_jaccard_idx}
\title{Jaccard similarity coefficient}
\usage{
calc_jaccard_idx(x, y)
}
\arguments{
\item{x}{A vector of positive reals}

\item{y}{A vector of positive reals}
}
\value{
A float: the Jaccard similarity coefficient
}
\description{
Calculate the Jaccard similarity coefficient
}
\details{
Considering that the classification produced by
}
\examples{
# this is an internal function, no example provided
}
\keyword{internal}
