% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_jaccard_mat}
\alias{calc_jaccard_mat}
\title{Jaccard similarity coefficient between columns of two matrices}
\usage{
calc_jaccard_mat(matX, matY)
}
\arguments{
\item{matX}{A matrix}

\item{matY}{A matrix}
}
\value{
A matrix with the Jaccard index values
}
\description{
Calculate the Jaccard similarity coefficient between the
columns of two matrices
}
\keyword{internal}
