% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_indices.R
\name{calc_local_moran_raster}
\alias{calc_local_moran_raster}
\title{Local Moran I for raster}
\usage{
calc_local_moran_raster(rast, window)
}
\arguments{
\item{rast}{A RasterLayer or a matrix}

\item{window}{The window defining the neighbour weights}
}
\value{
A RasterLayer or a matrix depending on the input with the local Moran I values
}
\description{
Calculate the Local Moran I for a numeric raster
}
\examples{
data("Arcachon")
rast <- Arcachon[[1]]
w <- matrix(1, nrow = 3, ncol = 3)
calc_local_moran_raster(rast, w)
}
