% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{calcBelongDenom}
\alias{calcBelongDenom}
\title{Intermediar step in the calculation of the belonging matrix}
\usage{
calcBelongDenom(centerid, alldistances, m)
}
\arguments{
\item{centerid}{An integer representing the column to use as reference in the
distance matrix}

\item{alldistances}{A distance matrix}
}
\value{
m a float representing the fuzzyness degree
}
\description{
Intermediar step in the calculation of the belonging matrix
}
\examples{
This is an internal function, no example provided
}
