% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcBelongMatrixNoisy}
\alias{calcBelongMatrixNoisy}
\title{Calculate the membership matrix with a noise cluster}
\usage{
calcBelongMatrixNoisy(centers, data, m, delta, sigmas)
}
\arguments{
\item{centers}{A matrix or a dataframe representing the centers of the
clusters with p columns and k rows}

\item{data}{A dataframe or matrix representing the observed data with n rows
and p columns}

\item{m}{A float representing the fuzziness degree}

\item{delta}{A float, the value set for delta by the user}

\item{sigmas}{A numeric vector for calculating the robust version of the FCM. Filled with ones
if the classical version is required}
}
\value{
A n * k matrix representing the probability of belonging of each
   observation to each cluster
}
\description{
Calculate the membership matrix according to a set of centroids, the observed
 data and the fuzziness degree
}
\keyword{internal}
