% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_evaluation.R
\name{calcDavidBouldin}
\alias{calcDavidBouldin}
\title{David-Bouldin index}
\usage{
calcDavidBouldin(data, belongmatrix, centers)
}
\arguments{
\item{data}{The original dataframe used for the clustering (n*p)}

\item{belongmatrix}{A membership matrix (n*k)}

\item{centers}{The centers of the clusters}
}
\value{
A float : the David-Bouldin index
}
\description{
Calculate the David-Bouldin index of clustering quality. This index
can be seen as a ratio between the within cluster dispersion and the
between cluster separation. A lower value indicates a higher cluster compacity
or a higher cluster separation.
}
\examples{
data(LyonIris)
AnalysisFields <-c("Lden","NO2","PM25","VegHautPrt","Pct0_14","Pct_65","Pct_Img",
"TxChom1564","Pct_brevet","NivVieMed")
dataset <- LyonIris@data[AnalysisFields]
queen <- spdep::poly2nb(LyonIris,queen=TRUE)
Wqueen <- spdep::nb2listw(queen,style="W")
result <- SFCMeans(dataset, Wqueen,k = 5, m = 1.5, alpha = 1.5, standardize = TRUE)
calcDavidBouldin(result$Data, result$Belongings, result$Centers)
}
