% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcEuclideanDistance2}
\alias{calcEuclideanDistance2}
\title{euclidean distance between rows of a matrix and a vector}
\usage{
calcEuclideanDistance2(y, x)
}
\arguments{
\item{y}{a matrix}

\item{x}{a vector (same length as ncol(matrix))}
}
\value{
a vector (same length as nrow(matrix))
}
\description{
euclidean distance between rows of a matrix and a vector
}
\keyword{internal}
