% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_indices.R
\name{calcFuzzyElsa_raster}
\alias{calcFuzzyElsa_raster}
\title{Local Fuzzy ELSA statistic for raster}
\usage{
calcFuzzyElsa_raster(rasters, window, matdist)
}
\arguments{
\item{rasters}{A List of RasterLayer or a List of matrices, or an array}

\item{window}{A binary (0,1) matrix representing the neighbours spatial weights when working
with rasters. The matrix must have odd dimensions.}

\item{matdist}{A matrix representing the dissimilarity between the clusters. The matrix must
be squared and the diagonal must be filled with zeros.}
}
\value{
A raster or a matrix (depending on the input): the values of
local fuzzy ELSA statistic
}
\description{
Calculate the Local Fuzzy ELSA statistic for a numeric raster
}
\examples{
# this is an internal function, no example provided
}
\keyword{internal}
