% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_evaluation.R
\name{calcGD53}
\alias{calcGD53}
\title{Generalized Dunn’s index (53)}
\usage{
calcGD53(data, belongmatrix, centers)
}
\arguments{
\item{data}{The original dataframe used for the clustering (n*p)}

\item{belongmatrix}{A membership matrix (n*k)}

\item{centers}{The centers of the clusters}
}
\value{
A float : the David-Bouldin index
}
\description{
Calculate the Generalized Dunn’s index (v53) of clustering quality.
This index is a ratio of the worst pair-wise separation of clusters and the
worst pair-wise compactness of clusters. Thus a higher value indicates a
better clustering.
}
\examples{
data(LyonIris)
AnalysisFields <-c("Lden","NO2","PM25","VegHautPrt","Pct0_14","Pct_65","Pct_Img",
"TxChom1564","Pct_brevet","NivVieMed")
dataset <- LyonIris@data[AnalysisFields]
queen <- spdep::poly2nb(LyonIris,queen=TRUE)
Wqueen <- spdep::nb2listw(queen,style="W")
result <- SFCMeans(dataset, Wqueen,k = 5, m = 1.5, alpha = 1.5, standardize = TRUE)
calcGD53(result$Data, result$Belongings, result$Centers)
}
