% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_evaluation.R
\name{calcNegentropyI}
\alias{calcNegentropyI}
\title{Negentropy Increment index}
\usage{
calcNegentropyI(data, belongmatrix, centers)
}
\arguments{
\item{data}{The original dataframe used for the clustering (n*p)}

\item{belongmatrix}{A membership matrix (n*k)}

\item{centers}{The centers of the clusters}
}
\value{
A float : the David-Bouldin index
}
\description{
Calculate the Negentropy Increment index of clustering quality.
This index based on the assumption that a normally shapped cluster is more
desirable. It uses the difference between the average negentropy
of all the clusters in the partition, and that of the  whole partition.
a smaller value indicates a better partition.
}
\examples{
data(LyonIris)
AnalysisFields <-c("Lden","NO2","PM25","VegHautPrt","Pct0_14","Pct_65","Pct_Img",
"TxChom1564","Pct_brevet","NivVieMed")
dataset <- LyonIris@data[AnalysisFields]
queen <- spdep::poly2nb(LyonIris,queen=TRUE)
Wqueen <- spdep::nb2listw(queen,style="W")
result <- SFCMeans(dataset, Wqueen,k = 5, m = 1.5, alpha = 1.5, standardize = TRUE)
calcNegentropyI(result$Data, result$Belongings, result$Centers)
}
