% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_evaluation.R
\name{calcQualIdx}
\alias{calcQualIdx}
\title{calculate the quality index required}
\usage{
calcQualIdx(name, ...)
}
\arguments{
\item{name}{The name of the index to calculate}

\item{...}{The parameters needed to calculate the index}
}
\value{
A float: the value of the index
}
\description{
A selector function to get the right quality index
}
\examples{
# this is an internal function, no example provided
}
\keyword{internal}
