% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcRobustSigmas}
\alias{calcRobustSigmas}
\title{Calculate sigmas for the robust version of the c-means algorithm}
\arguments{
\item{data}{A Numeric matrix representing the observed data with n rows
and p columns}

\item{belongmatrix}{A n X k matrix giving for each observation n, its
probability to belong to the cluster k}

\item{centers}{A c X k matrix giving for each cluster c, its center in k dimensions}

\item{m}{A float representing the fuzziness degree}
}
\value{
A vector with the sigmas for each cluster
}
\description{
Calculate sigmas for the robust version of the c-means algorithm
}
\keyword{internal}
