% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_evaluation.R
\name{calcSilhouetteIdx}
\alias{calcSilhouetteIdx}
\title{Fuzzy Silhouette index}
\usage{
calcSilhouetteIdx(data, belongings)
}
\arguments{
\item{data}{The original dataframe used for the clustering (n*p)}

\item{belongings}{A membership matrix (n*k)}
}
\value{
A float, the fuzzy Silhouette index
}
\description{
Calculate the Silhouette index of clustering quality.
}
\details{
The index is calculated with the function SIL.F from the package fclust.
When the dataset is too large, an approach by subsampling is used to avoid
crash.
}
