% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{calcSWFCCentroids}
\alias{calcSWFCCentroids}
\title{Calculate the new centroids of the clusters based on the belonging matrix (spatial version)}
\usage{
calcSWFCCentroids(data, wdata, belongmatrix, neighbours, m, alpha)
}
\arguments{
\item{data}{A dataframe or matrix representing the observed data with n rows
and p columns}

\item{wdata}{A dataframe or matrix representing the lagged observed data with
nrows and p columns}

\item{belongmatrix}{A n X k matrix giving for each observation n, its
probability to belong to the cluster k}

\item{neighbours}{A list.w object describing the neighbours typically
produced by the spdep package}

\item{m}{An integer representing the fuzzyness degree}

\item{alpha}{A float representing the weight of the space in the analysis (0
is a typical fuzzy-c-mean algorithm, 1 is balanced between the two
dimensions, 2 is twice the weight for space)}
}
\value{
A n X k matrix represening the belonging probabiblities of each
  datapoint to each cluster
}
\description{
Calculate the new centroids of the clusters based on the belonging matrix (spatial version)
}
\examples{
This is an internal function, no example provided
}
