% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_support.R
\name{calcWdataRaster}
\alias{calcWdataRaster}
\title{Calculate lagged values for a raster dataset}
\usage{
calcWdataRaster(w, dataset, fun, missing_pxl)
}
\arguments{
\item{w}{A matrix}

\item{dataset}{A list of rasters}

\item{fun}{A string giving the name of a function or a function or "nl"
for non-local method}

\item{missing_pxl}{A boolean vector of missing (FALSE) pixels}
}
\description{
Calculate lagged values for a raster dataset given a window
and an agregation function
}
\examples{
# this is an internal function, no example provided
}
\keyword{internal}
