% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_indices.R
\name{check_matdist}
\alias{check_matdist}
\title{Check validity of a dissimilarity matrix}
\usage{
check_matdist(matdist)
}
\arguments{
\item{matdist}{A dissimilarity matrix}
}
\description{
Check the validity of a dissimilarity matrix
}
\examples{
# this is an internal function, no example provided
}
\keyword{internal}
