% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_support.R
\name{check_raters_dims}
\alias{check_raters_dims}
\title{Check dimensions of a list of rasters}
\usage{
check_raters_dims(rasters)
}
\arguments{
\item{rasters}{A list of rasters}
}
\description{
Check if all the rasters in a list have the same dimensions
}
\examples{
# this is an internal function, no example provided
}
\keyword{internal}
