% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_support.R
\name{check_window}
\alias{check_window}
\title{Check the shape of a window}
\usage{
check_window(w)
}
\arguments{
\item{w}{A matrix}
}
\description{
Check is a window is squarred and have odd dimensions
}
\examples{
# this is an internal function, no example provided
}
\keyword{internal}
