% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp.R
\name{draw_boxplots}
\alias{draw_boxplots}
\title{Draw the box plots of pannel 3}
\usage{
draw_boxplots(dataset, variables, values)
}
\arguments{
\item{dataset}{The dataset used}

\item{variables}{The variables names used for clustering}

\item{values}{A boolean vector indicating which observations must
be put in red on the chart}
}
\description{
Draw the box plots of pannel 3
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
