% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp.R
\name{draw_byplot}
\alias{draw_byplot}
\title{Draw the scatter plot of pannel 2 and 3}
\usage{
draw_byplot(
  x,
  y,
  col_values,
  colors,
  gpcol,
  belongings,
  input,
  qual = FALSE,
  qual_colors = NULL
)
}
\arguments{
\item{x}{The values for the x axis}

\item{y}{The values for the y axis}

\item{col_values}{The values for the color}

\item{colors}{A vector with the colors to use}

\item{belongings}{the membership matrix}

\item{input}{the shiny app input}

\item{qual}{A boolean indicating if the plot must be drawn for pannel 2
(FALSE) or 3 (TRUE)}

\item{qual_colors}{The colors to use if qual = TRUE}

\item{gpvol}{The index of the color to use}
}
\description{
Draw the scatter plot of pannel 2 and 3
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
