% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Elsa_categorical_matrix_window}
\alias{Elsa_categorical_matrix_window}
\title{Elsa statistic calculated on a matrix with a given window}
\usage{
Elsa_categorical_matrix_window(mat, window, dist)
}
\arguments{
\item{mat}{an IntegerMatrix, must be filled with integer, -1 indicates NA values, categories must start at 0}

\item{window}{the window to use to define neighbours. 0 can be used to indicate that a cell is not a neighbour}

\item{dist}{a distance matrix between the categories}
}
\value{
a NumericVector : the local values of ELSA
}
\description{
method described here : https://doi.org/10.1016/j.spasta.2018.10.001
}
\keyword{internal}
