% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Elsa_fuzzy_matrix_window}
\alias{Elsa_fuzzy_matrix_window}
\title{Fuzzy Elsa statistic calculated on a matrix with a given window}
\usage{
Elsa_fuzzy_matrix_window(mats, window, dist)
}
\arguments{
\item{mats}{An array, each slice must contains the membership values of one group}

\item{window}{the window to use to define neighbours. 0 can be used to indicate that a cell is not a neighbour}

\item{dist}{a distance matrix between the groups}
}
\value{
a NumericVector : the local values of ELSA
}
\description{
This is an extension to the fuzzy classification case for the Elsa statistic
}
\keyword{internal}
