% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_indices.R
\name{elsa_fuzzy_vector}
\alias{elsa_fuzzy_vector}
\title{Local Fuzzy ELSA statistic for vector}
\usage{
elsa_fuzzy_vector(memberships, nblistw, matdist)
}
\arguments{
\item{memberships}{A membership matrix}

\item{matdist}{A matrix representing the dissimilarity between the clusters. The matrix must
be squared and the diagonal must be filled with zeros.}

\item{nbslitw}{The spatial weight matrix (nblistw object from spdep)}
}
\value{
A vector of local ELSA values
}
\description{
Calculate the Local Fuzzy ELSA statistic using a nblistw object
}
\examples{
# this is an internal function, no example provided
}
\keyword{internal}
