% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_indices.R
\name{elsa_raster}
\alias{elsa_raster}
\title{calculate ELSA spatial statistic for raster dataset}
\usage{
elsa_raster(rast, window, matdist)
}
\arguments{
\item{rast}{An integer raster or matrix representing the m categories (0,1,2,..., m)}

\item{window}{A binary (0,1) matrix representing the neighbours spatial weights when working
with rasters. The matrix must have odd dimensions.}

\item{matdist}{A matrix representing the dissimilarity between the clusters. The matrix must
be squared and the diagonal must be filled with zeros.}
}
\value{
A raster or a matrix: the local values of ELSA
}
\description{
calculate ELSA spatial statistic for vector dataset
}
\examples{
# this is an internal function, no example provided
}
\keyword{internal}
