% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_indices.R
\name{elsa_vector}
\alias{elsa_vector}
\title{calculate ELSA spatial statistic for vector dataset}
\usage{
elsa_vector(categories, nblistw, dist)
}
\arguments{
\item{categories}{An integer vector representing the m categories (1,2,3,..., m),
-1 is used to indicate missing values.}

\item{nblistw}{A listw object from spdep representing neighbour relations}

\item{dist}{A numeric matrix (m*m) representing the distances between categories}
}
\value{
A vector: the local values of ELSA
}
\description{
calculate ELSA spatial statistic for vector dataset
}
\examples{
# this is an internal function, no example provided
}
\keyword{internal}
