% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCMclass.R
\name{FCMres}
\alias{FCMres}
\title{Instantiate a FCMres object}
\usage{
FCMres(obj)
}
\arguments{
\item{obj}{A list, typically obtained from functions CMeans, GCMeans, SFCMeans, SGFCMeans}
}
\value{
A object of class FCMres
}
\description{
Instantiate a FCMres object from a list
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
