% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{focal_adj_mean_arr_window}
\alias{focal_adj_mean_arr_window}
\title{focal mean weighted by inverse of euclidean distance on a cube}
\usage{
focal_adj_mean_arr_window(mat, window)
}
\arguments{
\item{mat}{an array (cube)}

\item{window}{a numeric matrix (squared)}
}
\value{
a lagged version of the original cube
}
\description{
focal mean weighted by inverse of euclidean distance on a cube
}
\keyword{internal}
