% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{focal_euclidean_arr_window}
\alias{focal_euclidean_arr_window}
\title{focal euclidean distance on a matrix with a given window for a cube}
\usage{
focal_euclidean_arr_window(mat, window)
}
\arguments{
\item{mat}{an array (cube)}

\item{window}{a numeric matrix (squared)}
}
\value{
a matrix with the euclidean distance of each cell to
its neighbours.
}
\description{
focal euclidean distance on a matrix with a given window for a cube
}
\keyword{internal}
