% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{focal_euclidean_mat}
\alias{focal_euclidean_mat}
\title{focal euclidean distance on a matrix}
\usage{
focal_euclidean_mat(mat, w)
}
\arguments{
\item{mat}{a matrix}

\item{w}{an integer floor((nrow(window)/2))}
}
\value{
a matrix with the euclidean distance of each cell to
its neighbours.
}
\description{
focal euclidean distance on a matrix
}
\keyword{internal}
