% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{focal_euclidean_mat_window}
\alias{focal_euclidean_mat_window}
\title{focal euclidean distance on a matrix with a given window}
\usage{
focal_euclidean_mat_window(mat, window)
}
\arguments{
\item{mat}{a matrix}

\item{window}{a numeric matrix (squared)}
}
\value{
a matrix with the euclidean distance of each cell to
its neighbours.
}
\description{
focal euclidean distance on a matrix with a given window
}
\keyword{internal}
