% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_functions.R
\name{force_sp_sample}
\alias{force_sp_sample}
\title{wrapper around spsample}
\usage{
force_sp_sample(spatial, n, type)
}
\arguments{
\item{spatial}{An sp object}

\item{n}{The number of points}

\item{type}{The original strategy to use, changed for "regular" if the
original strategy fail}
}
\value{
a SpatialPoints object
}
\description{
A wrapper around spsample to avoid errors when points are
hard to find
}
\examples{
# this is an internal function, no example provided
}
\keyword{internal}
