% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocmeans.R
\docType{package}
\name{geocmeans-package}
\alias{geocmeans}
\alias{geocmeans-package}
\title{geocmeans: A package implementing methods for spatially constrained c-means
algorithm}
\description{
The geocmeans package implements a modified c-means algorithm more suited to
work with spatial data (characterized by spatial autocorrelation). The
spatial information is introduced with a spatial weight matrix W (n * n)
where wij indicate the strength of the spatial relationship between the
observations i and j. It is recommended to use a matrix standardized by row
(so that the sum of each row is 1). More specifically, the spatial c-means
combine the euclidean distance of each observation in the data matrix X to
each center with the euclidean distance of the lagged version of X by W (WX).
A parameter alpha controls for the weight of the lagged matrix. If
alpha = 0, then the spatial c-means is equal to a classical c-means. If alpha
= 1, then the weights given to  X and WX are equals. If alpha = 2, then the
weight of WX is twice the one of X and so on.
Several indices are provided to assess the quality of a classification on the
semantic and spatial dimensions. To explore results, a shiny app is also
available
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JeremyGelb/geocmeans}
  \item Report bugs at \url{https://github.com/JeremyGelb/geocmeans/issues}
}

}
\author{
\strong{Maintainer}: Jeremy Gelb \email{jeremy.gelb@ucs.inrs.ca} (\href{https://orcid.org/0000-0002-7114-2714}{ORCID})

Other contributors:
\itemize{
  \item Philippe Apparicio \email{philippe.apparicio@ucs.inrs.ca} (\href{https://orcid.org/0000-0001-6466-9342}{ORCID}) [contributor]
}

}
\keyword{internal}
