% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocmeans.R
\docType{package}
\name{geocmeans}
\alias{geocmeans}
\title{geocmeans: A package implementing methods for spatially constrained c-ceans algorithm}
\description{
The geocmeans, this section of the documentation is still to develop
}
\section{Foo functions}{

The foo functions ...
}

