% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp.R
\docType{data}
\name{geocmeans_env}
\alias{geocmeans_env}
\title{geocmeans general environment}
\format{
An object of class \code{environment} of length 0.
}
\usage{
geocmeans_env
}
\description{
An environment used by geocmeans to store data, functions and values
}
\keyword{internal}
