% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp.R
\name{hline}
\alias{hline}
\title{Add an horizontal line to plotly}
\usage{
hline(y = 0, color = "blue")
}
\arguments{
\item{y}{The y value for the line}

\item{color}{The color of the line}
}
\description{
Add an horizontal line to plotly
}
\examples{
#This is an internal function, no example provided
}
