% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_support.R
\name{input_raster_data}
\alias{input_raster_data}
\title{Raster data preparation}
\usage{
input_raster_data(dataset, w = NULL, fun = sum, standardize = TRUE)
}
\arguments{
\item{dataset}{A list of rasters}

\item{w}{The window to use in the focal function}

\item{fun}{the function to use as the focal function}

\item{standardize}{A boolean to specify if the variable must be centered and
reduced (default = True)}
}
\value{
A list with the required elements to perform clustering
}
\description{
Prepare a raster dataset
}
\examples{
# this is an internal function, no example provided
}
\keyword{internal}
