% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembling.R
\name{kppCenters}
\alias{kppCenters}
\title{kpp centers selection}
\usage{
kppCenters(data, k)
}
\arguments{
\item{data}{The dataset used in the classification}

\item{k}{The number of groups for the classification}
}
\value{
a DataFrame, each row is the center of a cluster
}
\description{
Select the inital centers of centroids by using the k++ approach
as suggested in this article: http://ilpubs.stanford.edu:8090/778/1/2006-13.pdf
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
