% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{local_moranI_matrix_window}
\alias{local_moranI_matrix_window}
\title{Local Moran I calculated on a matrix with a given window}
\usage{
local_moranI_matrix_window(mat, window)
}
\arguments{
\item{mat}{a matrix}

\item{window}{the window to use to define neighbours. 0 can be used to indicate that a cell is not a neighbour}
}
\value{
a double, the value of Moran I
}
\description{
Local Moran I calculated on a matrix with a given window
}
\keyword{internal}
