% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembling.R
\name{main_worker}
\alias{main_worker}
\title{Main worker function}
\usage{
main_worker(algo, ...)
}
\arguments{
\item{algo}{A string indicating the algorithm to use (one of FCM, GFCM, SGFCM)}

\item{...}{all the required arguments for the algorithm to use}
}
\value{
A named list with
\itemize{
        \item Centers: a dataframe describing the final centers of the groups
        \item Belongings: the final membership matrix
        \item Groups: a vector with the names of the most likely group for each observation
        \item Data: the dataset used to perform the clustering (might be standardized)
}
}
\description{
Execution of the classification algorithm
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
