% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_functions.R
\name{mapLines}
\alias{mapLines}
\title{Mapping the clusters (lines)}
\usage{
mapLines(geodata, belongmatrix, undecided = NULL)
}
\arguments{
\item{geodata}{A spatiallinesdataframe ordered like the original data used
for the clustering}

\item{belongmatrix}{The membership matrix obtained at the end of the algorithm}

\item{undecided}{A float between 0 and 1 giving the minimum value that an
observation must get in the membership matrix to not be considered as
uncertain (default = NULL)}
}
\value{
A named list with :
\itemize{
        \item ProbaMaps : a list of ggplot maps showing for each group the
        probability of the observations to belong to that group
        \item ClusterMap : a ggplot map showing the most likely group for each observation
}
}
\description{
Internal function to realize maps based on spatiallinesdataframe
}
\examples{
#No example provided, this is an internal function, use the general wrapper function mapClusters
}
\keyword{internal}
