% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{moranI_matrix}
\alias{moranI_matrix}
\title{Moran I calculated on a matrix}
\usage{
moranI_matrix(mat, w)
}
\arguments{
\item{mat}{a matrix}

\item{w}{the size of the neighbouring window}
}
\value{
a double, the value of Moran I
}
\description{
Moran I calculated on a matrix
}
\keyword{internal}
