% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_support.R
\name{output_raster_data}
\alias{output_raster_data}
\title{Raster result transformation}
\usage{
output_raster_data(object, missing, rst)
}
\arguments{
\item{object}{A FCMres object}

\item{missing}{A boolean indicating which pixels have no missing values}

\item{rst}{A raster object used as template to structure the results}
}
\value{
A FCMres object with isRaster = TRUE
}
\description{
Adapt the results if a raster is used
}
\examples{
# this is an internal function, no example provided
}
\keyword{internal}
