% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pow_matrices_bycol}
\alias{pow_matrices_bycol}
\alias{pow_matrix_bycol}
\title{element wise power of a matrix by column}
\usage{
pow_matrix_bycol(x, p)
}
\arguments{
\item{x}{a matrix}

\item{p}{the exponent}
}
\value{
a matrix
}
\description{
element wise power of a matrix by column
}
\keyword{internal}
