% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{power_mat}
\alias{power_mat}
\title{power of a matrix}
\usage{
power_mat(x, p)
}
\arguments{
\item{x}{a matrix}

\item{p}{a float}
}
\value{
x ** p
}
\description{
power of a matrix
}
\keyword{internal}
