% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_functions.R
\name{predict.FCMres}
\alias{predict.FCMres}
\title{Predict method for FCMres object}
\usage{
predict.FCMres(object, new_data, listw = NULL, standardize = TRUE, ...)
}
\arguments{
\item{object}{An FCMres object as produced by one of the following functions: CMeans,
GCMeans, SFCMeans, SGFCMeans}

\item{new_data}{A DataFrame with the new observations}

\item{listw}{A nb object from spdep (for the new data)}

\item{standardize}{A boolean to specify if the variable must be centered and
reduced (default = True)}

\item{...}{not used}
}
\value{
A numeric matrix with the membership values for each new observation
}
\description{
Function predict the membership matrix of a new set of observations
}
\examples{
data(LyonIris)
AnalysisFields <-c("Lden","NO2","PM25","VegHautPrt","Pct0_14","Pct_65","Pct_Img",
"TxChom1564","Pct_brevet","NivVieMed")

# rescaling all the variables used in the analysis
for (field in AnalysisFields) {
    LyonIris@data[[field]] <- scale(LyonIris@data[[field]])
}

# doing the initial clustering
dataset <- LyonIris@data[AnalysisFields]
queen <- spdep::poly2nb(LyonIris,queen=TRUE)
Wqueen <- spdep::nb2listw(queen,style="W")
result <- SGFCMeans(dataset, Wqueen,k = 5, m = 1.5, alpha = 1.5, beta = 0.5, standardize = FALSE)

# using a subset of the original dataframe as "new data"
new_data <- LyonIris[c(1, 27, 36, 44, 73),]
new_dataset <- new_data@data[AnalysisFields]
new_nb <- spdep::poly2nb(new_data,queen=TRUE)
new_Wqueen <- spdep::nb2listw(new_nb,style="W")

# doing the prediction
predictions <- predict(result, new_dataset, new_Wqueen, standardize = FALSE)
}
