% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prod_matrices_bycol}
\alias{prod_matrices_bycol}
\title{element wise product of two matrices by column}
\usage{
prod_matrices_bycol(x, y)
}
\arguments{
\item{x}{a matrix}

\item{y}{a matrix with the same dimensions}
}
\value{
a matrix
}
\description{
element wise product of two matrices by column
}
\keyword{internal}
