% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rowmins_mat}
\alias{rowmins_mat}
\title{minimum of each row of a matrix}
\usage{
rowmins_mat(x)
}
\arguments{
\item{x}{a matrix}
}
\value{
a NumericVector
}
\description{
minimum of each row of a matrix
}
\keyword{internal}
