% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembling.R
\name{sanity_check}
\alias{sanity_check}
\title{Parameter checking function}
\usage{
sanity_check(dots, data)
}
\arguments{
\item{dots}{A list of parameters used}

\item{data}{A numeric and complete dataframe}
}
\value{
A boolean, TRUE if all the tests are passed, FALSE otherwise
}
\description{
Check that the provided parameters are valid
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
