% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_functions.R
\name{select_parameters}
\alias{select_parameters}
\title{function to select the parameters of the classification alpha, k and m}
\usage{
select_parameters(
  data,
  k,
  m,
  alpha,
  nblistw,
  spconsist = T,
  classidx = T,
  standardize = T,
  maxiter = 500,
  tol = 0.01,
  seed = 123
)
}
\arguments{
\item{data}{a dataframe with numeric columns}

\item{k}{a sequence of values for k to test (>=2)}

\item{m}{a sequence of values for m to test}

\item{alpha}{a sequence of values for alpha to test}

\item{nblistw}{A list.w object describing the neighbours typically produced
by the spdep package}

\item{spconsist}{A boolean indicating if the spatial consistency must be
calculated}

\item{classidx}{A boolean indicating if the quality of classification
indices must be calculated}

\item{standardize}{A boolean to specify if the variable must be centered and
reduce (default = True)}

\item{maxiter}{An integer for the maximum number of iteration}

\item{tol}{The tolerance criterion used in the evaluateMatrices function for
convergence assessment}

\item{seed}{An integer used for random number generation. It ensures that the
start centers will be the same if the same integer is selected.}
}
\value{
a dataframe with indicators assessing the quality of classifications
}
\description{
function to select the parameters of the classification alpha, k and m
}
\examples{
data(LyonIris)
AnalysisFields <-c("Lden","NO2","PM25","VegHautPrt","Pct0_14","Pct_65","Pct_Img",
"TxChom1564","Pct_brevet","NivVieMed")
dataset <- LyonIris@data[AnalysisFields]
queen <- spdep::poly2nb(LyonIris,queen=TRUE)
Wqueen <- spdep::nb2listw(queen,style="W")
values <- select_parameters(dataset, k = 5, m = seq(2,3,0.1),
    alpha = seq(0,2,0.1), nblistw = Wqueen)
}
