% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp.R
\name{shiny_server}
\alias{shiny_server}
\title{Shiny App server}
\usage{
shiny_server(input, output, session)
}
\arguments{
\item{input}{The shiny input object}

\item{output}{The shiny output object}

\item{session}{The shiny session object}
}
\description{
Set the server for the Shiny App
}
\examples{
#This is an internal function, no example provided
}
\keyword{interal}
