% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp.R
\name{sp_clust_explorer}
\alias{sp_clust_explorer}
\title{Classification result explorer}
\usage{
sp_clust_explorer(belongings, dataset, spatial, ...)
}
\arguments{
\item{belongings}{A matrix or a dataframe representing the membership values
obtained for each observation}

\item{dataset}{A dataframe or matrix representing the data used for the
classification}

\item{spatial}{A spatial object (SpatialPointsDataFrame, SpatialPolygonsDataFrame or
SpatialLinesDataFrame) used to map the observations}

\item{...}{Other parameters passed to the function runApp}
}
\description{
Stat a local Shiny App to explore the results of a classification
}
\examples{
\dontrun{
data(LyonIris)

#selecting the columns for the analysis
AnalysisFields <-c("Lden","NO2","PM25","VegHautPrt","Pct0_14",
                   "Pct_65","Pct_Img","TxChom1564","Pct_brevet","NivVieMed")

#rescaling the columns
Data <- LyonIris@data[AnalysisFields]
for (Col in names(Data)){
  Data[[Col]] <- as.numeric(scale(Data[[Col]]))
}

Cmean <- CMeans(Data,4,1.5,500,standardize = FALSE, seed = 456, tol = 0.00001, verbose = FALSE)

sp_clust_explorer(Cmean$Belongings, Data, LyonIris)
}
}
