% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sqrt_matrix_bycol}
\alias{sqrt_matrix_bycol}
\title{element wise square root of a matrix by column}
\usage{
sqrt_matrix_bycol(x)
}
\arguments{
\item{x}{a matrix}
}
\value{
a matrix
}
\description{
element wise square root of a matrix by column
}
\keyword{internal}
