% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sub_matrices_bycol}
\alias{sub_matrices_bycol}
\title{substraction of two matrices by column}
\usage{
sub_matrices_bycol(x, y)
}
\arguments{
\item{x}{a matrix}

\item{y}{a matrix with the same dimensions}
}
\value{
a matrix
}
\description{
substraction of two matrices by column
}
\keyword{internal}
