% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_inferior_mat}
\alias{test_inferior_mat}
\title{create a logical matrix with inferior comparison}
\usage{
test_inferior_mat(mat, t)
}
\arguments{
\item{mat}{a matrix}

\item{t}{a double to compare}
}
\value{
a LogicalMatrix
}
\description{
create a logical matrix with inferior comparison
}
\keyword{internal}
