% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vector_out_prod}
\alias{vector_out_prod}
\title{create a matrix by multiplying a vector by its elements one by one as rows}
\usage{
vector_out_prod(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a NumericMatrix
}
\description{
create a matrix by multiplying a vector by its elements one by one as rows
}
\keyword{internal}
