% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp.R
\name{violinplots_ui}
\alias{violinplots_ui}
\title{Set the violin plot pannel}
\usage{
violinplots_ui(n, nr, nc)
}
\arguments{
\item{n}{The number of plots}

\item{nr}{The number of rows}

\item{nc}{The number of columns}
}
\description{
A simple function to prepare the violin plot pannel
}
\examples{
#This is an internal function, no example provided
}
