% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{adj_spconsist_arr_window_globstd}
\alias{adj_spconsist_arr_window_globstd}
\title{Adjusted spatial inconsistency index for rasters}
\usage{
adj_spconsist_arr_window_globstd(data, memberships, window)
}
\arguments{
\item{data}{an arma cube of dimension nr,nc,ns}

\item{memberships}{an arma cube of dimension nr, nc, ks}

\item{a}{matrix representing the neighbouring of each pixel}
}
\value{
a double, the adjusted spatial inconsitency index
}
\description{
Adjusted spatial inconsistency index for rasters
}
\keyword{internal}
