% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_jaccard_idx}
\alias{calc_jaccard_idx}
\title{Jaccard similarity coefficient}
\usage{
calc_jaccard_idx(x, y)
}
\arguments{
\item{x}{A vector of positive reals}

\item{y}{A vector of positive reals}
}
\value{
A double: the Jaccard similarity coefficient
}
\description{
Calculate the Jaccard similarity coefficient
}
\keyword{internal}
