% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_indices.R
\name{calc_moran_raster}
\alias{calc_moran_raster}
\title{Global Moran I for raster}
\usage{
calc_moran_raster(rast, window)
}
\arguments{
\item{rast}{A RasterLayer or a matrix}

\item{window}{The window defining the neighbour weights}
}
\value{
A float: the global Moran I
}
\description{
Calculate the global Moran I for a numeric raster
}
\examples{
data("Arcachon")
rast <- Arcachon[[1]]
w <- matrix(1, nrow = 3, ncol = 3)
calc_moran_raster(rast, w)
}
