% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_evaluation.R
\name{calc_raster_spinconsistency}
\alias{calc_raster_spinconsistency}
\title{calculate spatial inconsistency for raster}
\usage{
calc_raster_spinconsistency(matrices, w)
}
\arguments{
\item{matrices}{A list of matrices}

\item{w}{The window padding size}
}
\value{
A float: the sum of spatial inconsistency
}
\description{
Calculate the spatial inconsistenct sum for a set of rasters
}
\examples{
# this is an internal function, no example provided
}
\keyword{internal}
