% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_indices.R
\name{calc_raster_spinconsistency}
\alias{calc_raster_spinconsistency}
\title{calculate spatial inconsistency for raster}
\usage{
calc_raster_spinconsistency(matrices, window, adj = FALSE, dataset = NULL)
}
\arguments{
\item{matrices}{A list of matrices}

\item{window}{The window to use to define spatial neighbouring}

\item{adj}{A boolean indicating if the adjusted version of the algorithm must be
calculated}

\item{dataset}{A list of matrices with the original data (if adj = TRUE)}
}
\value{
A float: the sum of spatial inconsistency
}
\description{
Calculate the spatial inconsistency sum for a set of rasters
}
\examples{
# this is an internal function, no example provided
}
\keyword{internal}
