% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{calcBelongMatrix}
\alias{calcBelongMatrix}
\title{Calculate the belonging matrix according to a set of centroids, the observed
data and the fuzzyness degree}
\usage{
calcBelongMatrix(centers, data, m)
}
\arguments{
\item{centers}{A matrix or a dataframe representing the centers of the
clusters with p columns and k rows}

\item{data}{A dataframe or matrix representing the observed data with n rows
and p columns}

\item{m}{An float representing the fuzzyness degree}
}
\value{
A n * k matrix represening the probability of belonging of each
  datapoint to each cluster
}
\description{
Calculate the belonging matrix according to a set of centroids, the observed
data and the fuzzyness degree
}
\examples{
This is an internal function, no example provided
}
