% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{calcCentroids}
\alias{calcCentroids}
\title{Calculate the new centroids of the clusters based on the belonging matrix}
\usage{
calcCentroids(data, belongmatrix, m)
}
\arguments{
\item{data}{a dataframe or matrix representing the observed data with n rows
and p columns}

\item{belongmatrix}{a n X k matrix giving for each observation n, its
probability to belong to the cluster k}

\item{m}{an integer representing the fuzziness degree}
}
\value{
a n X k matrix representing the belonging probabilities of each
  observation to each cluster
}
\description{
Calculate the new centroids of the clusters based on the belonging matrix
}
\examples{
#This is an internal function, no example provided
}
