% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCM.R
\name{calcCentroids}
\alias{calcCentroids}
\title{Calculate the centroids}
\usage{
calcCentroids(data, belongmatrix, m)
}
\arguments{
\item{data}{A dataframe or matrix representing the observed data with n rows
and p columns}

\item{belongmatrix}{A n X k matrix giving for each observation n, its
probability to belong to the cluster k}

\item{m}{An integer representing the fuzziness degree}
}
\value{
A n X k matrix representing the belonging probabilities of each
  observation to each cluster
}
\description{
Calculate the new centroids of the clusters based on the membership matrix
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
