% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembling.R
\name{calcEuclideanDistance}
\alias{calcEuclideanDistance}
\title{Calculate the Euclidean distance}
\usage{
calcEuclideanDistance(m, v)
}
\arguments{
\item{m}{A n * p matrix or dataframe with only numeric columns}

\item{v}{A numeric vector of length p}
}
\value{
A vector of length n giving the euclidean distance between all matrix
  row and the vector p
}
\description{
Calculate the euclidean distance between a numeric matrix n * p and a numeric
vector of length p
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
