% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembling.R
\name{calcLaggedData}
\alias{calcLaggedData}
\title{Lagged Data}
\usage{
calcLaggedData(x, nblistw, method = "mean")
}
\arguments{
\item{x}{A dataframe with only numeric columns}

\item{nblistw}{The listw object (spdep like) used to calculate WY}

\item{method}{A string indicating if a classical lag must be used
("mean") or if a weighted median must be used ("median")}
}
\value{
A lagged version of x
}
\description{
Calculate Wx, the spatially lagged version of x, by a neighbouring matrix W.
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
