% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{calcSFCMBelongMatrix}
\alias{calcSFCMBelongMatrix}
\title{Calculate the belonging matrix (spatial version) according to a set of
centroids, the observed data, the fuzzyness degree a neighbouring matrix and
a spatial ponderation term}
\usage{
calcSFCMBelongMatrix(centers, data, wdata, m, alpha)
}
\arguments{
\item{centers}{A matrix or a dataframe representing the centers of the
clusters with p columns and k rows}

\item{data}{A dataframe or matrix representing the observed data with n rows
and p columns}

\item{wdata}{A dataframe or matrix representing the lagged observed data with
nrows and p columns}

\item{m}{A float representing the fuzzyness degree}

\item{alpha}{A float representing the weight of the space in the analysis (0
is a typical fuzzy-c-mean algorithm, 1 is balanced between the two
dimensions, 2 is twice the weight for space)}
}
\value{
A n * k matrix represening the belonging probabilities of each
  datapoint to each cluster
}
\description{
Calculate the belonging matrix (spatial version) according to a set of
centroids, the observed data, the fuzzyness degree a neighbouring matrix and
a spatial ponderation term
}
\examples{
#This is an internal function, no example provided
}
