% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcSFCMBelongMatrix}
\alias{calcSFCMBelongMatrix}
\title{Calculate the membership matrix (spatial version)}
\usage{
calcSFCMBelongMatrix(centers, data, wdata, m, alpha)
}
\arguments{
\item{centers}{A matrix or a dataframe representing the centers of the
clusters with p columns and k rows}

\item{data}{A matrix representing the observed data with n rows and p columns}

\item{wdata}{A matrix representing the lagged observed data with n rows and p columns}

\item{m}{A float representing the fuzziness degree}

\item{alpha}{A float representing the weight of the space in the analysis (0
is a typical fuzzy-c-mean algorithm, 1 is balanced between the two
dimensions, 2 is twice the weight for space)}
}
\value{
A n * k matrix representing the belonging probabilities of each
  observation to each cluster
}
\description{
Calculate the membership matrix (spatial version) according to a set of
centroids, the observed data, the fuzziness degree a neighbouring matrix and
a spatial weighting term
}
\keyword{internal}
