% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_functions.R
\name{describGroups}
\alias{describGroups}
\title{display some descriptive informations about fixed groups}
\usage{
describGroups(data, groupvar, vars, dec = 5)
}
\arguments{
\item{data}{a dataframe with numeric columns}

\item{groupvar}{the name of the categorical variable used to split the
dataset}

\item{vars}{the names of the numerical variables}

\item{dec}{the number of digits to keep}
}
\value{
a named list with :
\itemize{
        \item Violin : a list with a length equal to the number  of columns in data.
        Each element of the list in a violin plot usefull to compare the distributions
        of each group on that variable
        \item Tableau : a list with a length equal to the number of group (k). Each
        element in a dataframe with summary statistics describing the group for the
        columns of data.
}
}
\description{
display some descriptive informations about fixed groups
}
\examples{
data(LyonIris)
AnalysisFields <-c("Lden","NO2","PM25","VegHautPrt","Pct0_14","Pct_65","Pct_Img",
"TxChom1564","Pct_brevet","NivVieMed")
dataset <- LyonIris@data[AnalysisFields]
queen <- spdep::poly2nb(LyonIris,queen=TRUE)
Wqueen <- spdep::nb2listw(queen,style="W")
result <- SFCMeans(dataset, Wqueen,k = 5, m = 1.5, alpha = 1.5, standardize = TRUE)
dataset$group <- result$Groups
describGroups(dataset, "group", AnalysisFields)
}
