% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_functions.R
\name{eval_parameters}
\alias{eval_parameters}
\title{Worker function for select_parameters and select_parameters.mc}
\usage{
eval_parameters(
  parameters,
  data,
  nblistw,
  standardize,
  spconsist,
  classidx,
  tol,
  maxiter,
  seed = 123
)
}
\arguments{
\item{parameters}{a dataframe of parameters with columns k,m and alpha}

\item{data}{a dataframe with numeric columns}

\item{nblistw}{a list.w object describing the neighbours typically produced
by the spdep package}

\item{standardize}{a boolean to specify if the variable must be centered and
reduce (default = True)}

\item{spconsist}{a boolean indicating if the spatial consistency must be
calculated}

\item{classidx}{a boolean indicating if the quality of classification
indices must be calculated}

\item{tol}{the tolerance criterion used in the evaluateMatrices function for
convergence assessment}

\item{maxiter}{an integer for the maximum number of iteration}

\item{seed}{an integer used for random number generation. It ensures that the
start centers will be the same if the same integer is selected.}
}
\description{
Worker function for select_parameters and select_parameters.mc
}
\examples{
#No example provided, this is an internal function
}
