% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembling.R
\name{evaluateMatrices}
\alias{evaluateMatrices}
\title{Matrix evaluation}
\usage{
evaluateMatrices(mat1, mat2, tol)
}
\arguments{
\item{mat1}{A n X k matrix giving for each observation n, its probability to
belong to the cluster k at iteration i}

\item{mat2}{A n X k matrix giving for each observation n, its probability to
belong to the cluster k at iteration i+1}

\item{tol}{A float representing the algorithm tolerance}
}
\value{
A boolean, TRUE if the test is passed, FALSE otherwise
}
\description{
Evaluate if the algorithm converged by comparing two successive membership
matrices. Calculate the absolute difference between the matrices and then
calculate the max of each row. If all the values of the final vector are
below the fixed tolerance, then return True, else return False
}
\examples{
#This is an internal function, no example provided
}
\keyword{internal}
