% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{focal_euclidean}
\alias{focal_euclidean}
\alias{focal_euclidean_list}
\title{focal euclidean distance on a list of matrices}
\usage{
focal_euclidean_list(matrices, window)
}
\arguments{
\item{matrices}{a List of matrices with the same dimensions}

\item{window}{a numeric matrix}
}
\value{
a matrix with the euclidean distance of each cell to
its neighbours.
}
\description{
focal euclidean distance on a list of matrices
}
\keyword{internal}
