% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_functions.R
\name{select_parameters.mc}
\alias{select_parameters.mc}
\title{Function to select the parameters of the classification alpha, k and m.
This version of the function allows to use a plan defined with the package
future to reduce calculation time}
\usage{
select_parameters.mc(
  data,
  k,
  m,
  alpha,
  nblistw,
  lag_method = "mean",
  spconsist = T,
  classidx = T,
  standardize = T,
  maxiter = 500,
  tol = 0.01,
  seed = 123,
  chunk_size = 100
)
}
\arguments{
\item{data}{a dataframe with numeric columns}

\item{k}{a sequence of values for k to test (>=2)}

\item{m}{a sequence of values for m to test}

\item{alpha}{a sequence of values for alpha to test}

\item{nblistw}{a list of list.w objects describing the neighbours typically
produced by the spdep package}

\item{lag_method}{a string indicating if a classical lag must be used
("mean") or if a weighted median must be used ("median"). Both can be
tested by specifying a vector : c("mean","median")}

\item{spconsist}{a boolean indicating if the spatial consistency must be
calculated}

\item{classidx}{a boolean indicating if the quality of classification
indices must be calculated}

\item{standardize}{a boolean to specify if the variable must be centered and
reduce (default = True)}

\item{maxiter}{an integer for the maximum number of iteration}

\item{tol}{The tolerance criterion used in the evaluateMatrices function for
convergence assessment}

\item{seed}{an integer used for random number generation. It ensures that the
start centers will be the same if the same integer is selected.}

\item{chunk_size}{The size of a chunk used for multiprocessing. Default is 100.}
}
\value{
a dataframe with indicators assessing the quality of classifications
}
\description{
Function to select the parameters of the classification alpha, k and m.
This version of the function allows to use a plan defined with the package
future to reduce calculation time
}
\examples{
data(LyonIris)
AnalysisFields <-c("Lden","NO2","PM25","VegHautPrt","Pct0_14","Pct_65","Pct_Img",
"TxChom1564","Pct_brevet","NivVieMed")
dataset <- LyonIris@data[AnalysisFields]
queen <- spdep::poly2nb(LyonIris,queen=TRUE)
Wqueen <- spdep::nb2listw(queen,style="W")
future::plan(future::multiprocess(workers=2))
#values <- select_parameters.mc(dataset, k = 2:5, m = seq(2,3.5,0.1),
#    alpha = seq(0,2,0.1), nblistw = Wqueen, chunk_size=50)
values <- select_parameters.mc(dataset, k = 5, m = seq(2,3,0.1),
    alpha = seq(0,2,0.1), nblistw = Wqueen)
\dontshow{
   ## R CMD check: make sure any open connections are closed afterward
   if (!inherits(future::plan(), "sequential")) future::plan(future::sequential)
}
}
